requirements.txt
requests
beautifulsoup4
emoji
python-dotenv
